Refactor: Change jwt package hierarchy
