+++ src/main/java/com/petspace/dev/service/UserService.java
@@ -8,8 +8,8 @@ import com.petspace.dev.dto.user.UserJoinRequestDto;
 import com.petspace.dev.dto.user.UserLoginRequestDto;
 import com.petspace.dev.dto.user.UserResponseDto;
 import com.petspace.dev.repository.UserRepository;
-import com.petspace.dev.service.auth.JwtProvider;
-import com.petspace.dev.service.auth.jwt.Token;
+import com.petspace.dev.util.jwt.JwtProvider;
+import com.petspace.dev.util.jwt.Token;
 import com.petspace.dev.util.exception.ReissueException;
 import com.petspace.dev.util.exception.UserException;
 import lombok.RequiredArgsConstructor;
