+++ src/main/java/com/petspace/dev/service/auth/OauthService.java
@@ -6,7 +6,8 @@ import com.petspace.dev.dto.auth.LoginTokenResponseDto;
 import com.petspace.dev.dto.auth.OauthRequestDto;
 import com.petspace.dev.repository.UserRepository;
 import com.petspace.dev.service.RedisService;
-import com.petspace.dev.service.auth.jwt.Token;
+import com.petspace.dev.util.jwt.JwtProvider;
+import com.petspace.dev.util.jwt.Token;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.core.ParameterizedTypeReference;
